Created a project with a simple form with a combobox, two buttons and a textbox.
Created a project with Visual Studio that contains a simple form with  a combobox, two buttons and a textbox.
When you click on "Even (par)", charge even number from 0 to 10 on the combobox.
When you click on "Odd (impar)", charge odd number from 0 to 10 on the combobox.
And when you select a number on the comboboxt, the number is copied into the textbox.